+++ src/router/index.ts
@@ -1,8 +1,5 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import HomeView from '../views/HomeView.vue'
-import SecondAssessment from '@/views/SecondAssessment.vue'
-import FirstAssessmentPage from '@/views/FirstAssessmentPage.vue'
-import ResultScreenView from '@/views/ResultScreenView.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -12,20 +9,8 @@ const router = createRouter({
       redirect: '/home'
     },
     {
-      path: '/second-assessment',
-      component: SecondAssessment
-    },
-    {
       path: '/home',
       component: HomeView
-    },
-    {
-      path: '/result',
-      component: ResultScreenView
-    },
-    {
-      path: '/assessment-page',
-      component: FirstAssessmentPage
     }
   ]
 })
